feat: download img, upload to twitter, remove local file
The key to deleting file is to add callback(err) in T.post. This passes error along to callback. It seems that this is the only way to move along to another function after T.post (or any twit function) has been called. In the postTweet callback, I passed along err as a parameter and then ran an if statement. If no error, then run deleteImage().